pydpwte/requirements.txt
setuptools==49.2.0
pytorch
lifelines
scipy==1.3.0
scikit-learn==0.21.2
numpy==1.16.4
progressbar2==3.50.1
pandas==1.1.1